Update Director to use new origin info
The Director must be able to redirect to origins, so we modified
topology such that the needed information can be fed into the
Director. Now the `/api/v1.0/director/object` and
`/api/v1.0/director/origin` endpoints direct to caches and origins,
respectively.

There are also a few smaller bug fixes in here, like fixing the
ttlcache timing out and having no reload mechanism.