Embed prometheus TSDB and scrape into web UI
This pulls the prometheus monitoring stack into the web UI allowing
any future React-like Javascript application to have a self-contained
way for basic monitoring data.  The vision is this will power visualizations
in the admin dashboard.

For simplicity of review, the changes to go.mod/go.sum will come in a
follow-up commit.