Fix bad error message (bad content length) (#10)
The error message 'bad content length' was given when the local copy of
a file is larger than the remote server copy when downloading a file.
The error message now reads 'Local copy of file is larger than remote
copy' which I believe is much clearer. In addition, fixed the bug
stating the error is retryable. It no longer states this type of error
is retryable.