Record stack trace of panic at the debug level
Previously, the panic was logged at the Error level; this wasn't useful
as the string gets printed out twice to stderr (as the panic is converted
to an error and re-raised).  With this, clients that have debugging
enabled will have far more useful information on panics for the dev team.