Use correct permissions when run as root.
Where possible, if run by root, the mode for created directories
should be 0750 with the ownership of root:xrootd.  This means that
the root user can change the resulting files but xrootd can still
read it.

A notable exception is the TLS key file which xrootd rejects if it
is group-readable.