Create a single `--port` flag object
When a flag is bound to the viper key (such as binding the `--port`
flag object to the `WebPort` key), it's a global.  If you do this
multiple times with multiple pointers, only the last one is used.

That means, if we invoke `cmd.Flags().StringP(...)` in multiple sub-commands,
only one will actually set `WebPort`.

The solution for this is to create the flag once globally, then add
the global object to multiple commands.  Note that I didn't use a
parent-level "persistent flag" as that would also make `--port` usage
strings appear for the `object` sub-command (which doesn't have an
embedded webserver).