Incorporate feedback from review
I also made the decision to enforce that the jwks used to register a namespace
prefix only contain one key. This simplifies things at both the client
and at the registry, because we need not figure out which key was used for
the handshake signature, and we need not support multiple signatures during
this part of the process. Eventually we need to expose an API that allows
origins to add/verify other public keys to the registry's jwks database.