Use full namespace object, not stub
When generating the origin record, we need the full namespace object
filled out (including the authorization and token-generation info).
This is because the origin's policy is used when later responding to
clients.

The same thing is not needed for the cache ads; the cache namespace
info is only used for determining whether or not a cache supports
a particular namespace.