Fix bug with common transport TLSSkipVerify
I found a bug with setting up TLSSkipVerify from the common transports.
Looks like I put checking for that within the creation of the transport
that only happens once. Therefore, if the transport is created, then it
will never check again if TLSSkipVerify is set. From running my own
tests by hand it seems to be working now.