Fix token verification
With this commit, almost all of the pieces for an origin to advertise directly
to the namespace registry are in place. A few final things should be tidied up,
but the basic functionality now exists. Before we can call this "done for now",
we need to look over line 60/61 of `origin_ui/advertise.go`. There's an
important TODO there that it looks like someone else has sitting in a branch.

Cleanup linter errors

Linter's stern decree,
Merge awaits its nod, unsure,
PR in limbo