Remove hardcoded originUrl
This commit removes the hardcoded origin url and instead makes the value configurable
via the pelican config file. Specifying an origin url is required to serve an origin,
and the command `pelican origin serve` will fail unless the value is set. It will also
fail if the configured origin url specifies a port that doesn't match the port passed
via the `-p` flag.