Updated mpb bar progress bars
This fixes a bug when downloading from an origin for the first time and a weird error message shows up. Turns out the package needed some updating. This took some time since
the function we originally had for v7 no longer existed in v8. Also changed around the styling a tad and made it look nice and work with this new version upgrade. Removed what
I thought was an unnecessary extra bar being made to say "done". Also found a typo to be fixed within handle_http