Updated progress bars
Updated progress bars per requests from PR. Made measuring progress actually work now (wasn't before) by calculating correct time and bytes that have progressed. Fixed a few leftover typos and changed up error messages to appear cleaner and give better worded info.

The only issue: I could not get a logrus warning message to print with the progress bars. The bar would quickly override the message and would ususally end up outputting 2 bars instead of 1.
My current fix: There is a function in the mpb package that allows you to write above the progress bars (but has to be a byte string). I could not figure out how to convert a logrus warning to a byte string to be outputted (if there even is a way). Therefore, I created a little warning to appear over the progress bar if the download is happening too slow. Even though it is not a logrus warning, I feel still helpful to know.