Add unit test and fix second layer of bug
Unsurprisingly, it turns out writing the unit test was a great idea because it caught
an additional bug that was a bit more nuanced. This time, the bug came from the fact
that some caches/origins might support a namespace prefix that's shorter than the best/
longest prefix without supporting that longest prefix. In those cases, the wrong
namespace (and slices of origins/caches) might wind up being returned not by
`matchesPrefix`, but by `GetCacheAdsForPath` due to the fact that `/chtc` IS the best
namespace prefix for the path `/chtc/PUBLIC/foo` if a cache only supports `/chtc`.

I don't love all having to deal with all of these trailing /'s, but because I'm assuming
there's a reason that was allowed in topology, so I'm hoping my test cases are robust enough
to catch all the nuanced edge cases.