Expose CLI for creating tokens
This PR exposes the ability for a user at an origin to create tokens for testing. It can
be accessed via `pelican origin token create`. There are several claims that all default
tokens will posses, namely iat, exp, nbf, iss, and jti. These can be overridden by passing
them specifically to the tool as claims.

By passing a profile with the --profile flag, the tool will provide information about
additional claims that are needed by the profile. NOTE: it only provides information about
required values that ARE NOT auto-generated!

The lifetime and private-key flags can be used to set the token lifetime (in seconds) and
the private signing key, respectively.

Next up is to create a CLI for verifying a provided token against a specific profile. While
the basic infrastructure was set up for that via `pelican origin token verify`, the function
remains unimplemented.