Add self-tests for the origin
With this, the origin will periodically create a token, upload a file,
download a file, and then delete the file from a special staging directory.

Includes the automatic configuration for the Authfile, a token issuer,
and the scitokens config.

Changes the automatic host certificate generation to utilize the generated
CA.  This is required for XRootD authorization to work from the SciTokens
library.