Add director min version detection capability
According to issue #21, users have encountered issues with stashcp getting confusing
error messages from the Director. While this PR doesn't necessarily handle old stashcp
clients that don't set user-agent headers when querying the director, it makes things
more robust moving forward.

Each service that queries the director now sets a User-Agent header of the form
"pelican-<service>/<version>", eg "pelican-client/7.0.0". The director checks if the
UA string matches the form, and if it doesn't it continues (this is to allow unknonwn
UAs to continue working, for example if someone wants to curl the director). If the
UA matches the expected form, it will parse both the service and the version and make
sure that it believes it can serve a service with that version.

For now, I set those minimum versions to 7.0.0 for both the client and the origin.