Strip "Bearer " from authz query and header if they're present
I noticed while testing XRootD's ability to use a Pelican director as its
redirector (defined via the `pss.origin` directive), that XRootD will take
a token passed to it via an Authorization header and convert it directly into
an authz URL query that gets passed to the director. For the director to return
a valid redirect URL, the "Bearer " prefix must be stripped.