Update Pelican to not require a set $HOME in some configurations
In some remote environments like HTCondor, we can't always assume that the user
has a configured $HOME directory. Previously, this was enforced in both the root
command and in the `main` package's chain of `init()` functions, preventing
stash_plugin from running in HTCondor even though no $HOME directory is needed
in that case.

This makes an unset $HOME a non-fatal warning at the root command level, and moves
`InitServer()` into a pre-run for each cobra command that is actually responsible
for starting a server. In the cases of servers, we do still require $HOME.