Cleanup of pelican config files
Fixes issues with the 3 config files used with Pelican: defaults.yaml, osdf.yaml, and pelican.yaml. Before this fix there was a bit of a mess of where the configs are set up with viper. This fix moves InitConfig to config.go and sets up all the config files in the priority they need to be with pelican.yaml taking higher precedence than defaults.yaml. Also fixed issue in object_copy.go with some weird issues with the -d flag. Only need to check for the -d flag with stashcp since it does not go through root, any other modes are checked within root/initconfig and trigger debug mode.