Remove access log config from tomcat
The access log is effectively duplicate information as the same thing
is recorded by the gin logger as well.  Further, when run as root within
a docker container, there an ugly stack trace about permission denied when
opening `/dev/stdout`.

In the end, we don't need it so remove it.