Fixed adding log levels to config
Now that the configurations are cleaned up, can successfully add the log
level as part of config. This should work for all methods (pelican,
osdf, and stashcp). However, for stashcp, nothing happens for log levels
set in config and you must use the `-d` flag since it does not go
through root.