Make director service host-aware
In some cases, the director service should be aware of the originally-requested host.
This is useful in the case where a host `https://director-origins.osg-htc.org` contains
a CNAME record to `https://director-caches.osg-htc.org`, because this indicates to the
cache director that it should actually redirect to an origin.

This works because the host header from the original request to the origin director is
forwarded through the CNAME redirect instead of being swapped out for the canonical host
from DNS. This essentially makes the director aware of who the request was originally
intended for, thus allowing it to act appropriately on the original request.

For now, if the Host or X-Forwarded-Host headers contain the word "origin", a cache director
will redirect to an origin, and vice-versa for an origin director. We may eventually need to
make these checks more intelligent. Note that the X-Forwarded-Host header is checked because
of the way we deploy services on Kubernetes through Traefic