Add pelican-osdf-compat package
This adds compat symlinks for `/usr/bin/stashcp`, `/usr/bin/osdf`, and
`condor-stash-plugin` to use the `pelican` binary isntead, and a condor
configuration for enabling the condor-stash-plugin.  This allows pelican
to replace the old "stashcp" and "condor-stash-plugin" RPMs.

The appropriate `obsoletes` (`replaces` for Deb) and `provides` clauses
are included, with the caveat that pelican-osdf-compat provides version
"7" of the packages it's replacing, as opposed to whatever specific
version the pelican-osdf-compat package itself is.  This is because
GoReleaser doesn't expand templates inside `replaces` and `provides`
lists so the value has to be hardcoded.

Deb, RPM, and APK use different syntax for specifying versions clauses:
- APK does not use `replaces` (`provides` is sufficient for replacing a package).
- Deb puts parens around the version restriction (so `stashcp (= 7)`
  whereas RPM and APK use `stashcp = 7`)
- Deb uses `<<` instead of `<` as the strictly-less-than operator.

The APK doesn't have a symlink and config file for condor-stash-plugin
because condor is not available on Alpine.

See #246