Fix null dereference bug in Director
The Director had a null dereference bug that occurred when it tried to find
ads for non-existent namespace prefixes. This commit fixes the bug, adds a
regression test to catch similar bugs in the future, and re-orders the error
reporting to the client in the event that it tries to get something from a
non-existent namespace.