Automatically redirect caches to origins
When a cache uses the director as its redirector service, it can be redirected to
one of two things: a cache or an origin. While it may likely be beneficial someday
to chain caches together (eg tell this cache to get the object from another cache),
that isn't feasible at the moment. As such, it only makes sense to redirect caches
to origins.

This commit has the director aggregate all of the cache server URLs that come from
topology and add them to the list of hostnames that should automatically be redirected
to origins. Eventually, when we have automatically-registered caches, those should
probably also implement this logic so that they can be correctly redirected as well.