Standardize server init and cleanup a few other small items
The big part of this PR is standardizing some of how servers are initialized. Importantly,
this moves federation discovery and some key creation from the various `*_serve.go` files
into `InitServer()`.

In the process of putting this together, I noticed a few other cleanup items. One was changing
a parameter.yaml default value and type to match the description. Another was modifying config_test
to use `assert.Equal` instead of `assert.True` so that if the test fails, it prints out what the
expected and actual values are.