Make Linux CA Certificate detection more robust and fix a few other small things
The big thing with this is to use the same list of default CA Certificate directories used
by the x509 package when trying to determine whether there's a default certificate on whatever
linux system we happen to be using. This arose out of the fact that we hard-coded the default
value for RHEL-derived linux, which started causing test problems in our Ubuntu test driver.

If no default certificate is found, we create one in $configDir/certificates (where we know we
can write to) and we then make sure to use that one when initializing our http.Transport object —
this also required running setupTransport() **after** we create the certificate, which we didn’t
always do before.

In addition to that, this also removes a few more redundant entries in namespace_registry_serve.go
and re-adds the GenerateIssuerJWKS() function to InitServer. While I caught that we needed to generate
the key for that earlier, I missed that generating the JWKS was a separate step, and Haoming pointed
out it is an important one 😉