Avoid narrowing issues around UID/GIDs
On Unix, the UID/GID type is limited to 32-bit unsigned integers.
Ensure, when we parse out the UID/GIDs from the go runtime, the
resulting values fit into the valid value range and there's not
an overflow when converting to syscalls that take a uint32.

This issue was detected by the CodeQL scanner.