Fix small bugs in DiscoverFederation()
Previously, running pelican with `-f foo.com/a/path` would result in DiscoverFederation()
trying to get metadata from `foo.com/.well-known/pelican-configuration/a/path` instead of
the correct `foo.com/a/path/.well-known/pelican-configuration`. This fixes that.

It also uses the `url.JoinPath()` method for parsing URL paths instead of `path.Join()`.
We never really do much with Windows, but if we did, that would prevent federation discovery
from working, because it would try to use \ as a path separator in a URL. Whoops.