Make Pelican able to use S3 xrootd plugin for unauthenticated S3 endpoints
This is the first step in giving Pelican the ability to use an S3 backend. For now, it only covers
unauthenticated buckets. Importantly, this doesn't do anything to actually make the plugin available
to xrootd unless its installed by hand. That will come in a later commit.

Add plugin to containers

Right now, the RPM for the S3/HTTP plugin isn't functional, and we don't have that repo's
GH set up to publish anything automatically. Until those can be sorted out, we'll clone from
source to install. Concurrent to this commit, I'm opening a few issues at the plugin repo
to handle getting the various GHA set up to take care of this.