Add boilerplate to allow an authenticated bucket
This adds what we need to set an access/secret key for the S3 plugin. HOWEVER,
upon trying to test this locally with MinIO, I found some bugs in the actual plugin
that prevent authenticated buckets from working in the first place. That being said,
that's a problem with the plugin and not with Pelican. All Pelican needs to do is to
run xrootd with the right `s3` directives, and that appears to be the case with these
changes.