Add very basic test for S3 origin and add dependency to dev.Dockerfile
This adds a very basic test for the S3 origin -- basically it only checks that
we can start an xrootd S3 server with the provided configuration. The reason for
this punt on the E2E testing is that I'm waiting for things like `originServe` to be
unwrapped from the cobra cli, because otherwise there's an awful lot of setup to do.

Since I don't actually test a file download, there's technically no need to set up the
MinIO server in the test, but that's there and ready when those other pieces are
figured out.

The last bit is that for these tests to run, the minio server needs to be installed. With
changes to `dev.Dockerfile`, the tests should be runnable.