Enable key chaining in the registry
Let's say that `/foo/bar` is already registered with some set of associated public keys. It
makes sense that we wouldn't want someone else to register `/foo` or `/foo/bar/baz` unless
they had the permission of the owner of `/foo/bar`, because we treat namespaces hierarchically.
This token chaining feature allows a registry to be run such that it will deny requests to
register `/foo` or `/foo/bar/baz` unless the registration token passed during registration is
signed by a key that's already associated with `/foo/bar`.