Add topology checks to key chaining
If key chaining is on, then we refuse to add super/subspaces that exist in topology.
The reason is that, for now, there are no public keys associated with those namespaces,
and without a public key, we have no way to verify the legitimacy of the incoming
registration request.

If token chaining is off, we still prevent re-registering in the registry any prefix
already registered in topology -- but we don't prevent registration of super/subspaces.