Allow GeoIP overrides and fix server geoip randomization
First and foremost, this allows a director to configure IP blocks that we want to
directly map to a lat/long pair. Valid IPs include IPv4 and IPv6, both masked and
unmasked. In the event that MaxMind either doesn't contain a coordinate for an IP
or we just don't like the coordinate it returns, we can now set where we want a
director to think the server is. Note that this affects both servers (origins/caches) as
well as clients, because we update server coordinates at ad recording time, and
we get client coordinates at redirect time.

One small change is that we now log at the Director when an IP address results in the
null coordinate. This should allow us to filter such requests in the logs when we're
collecting metrics.

This also fixes another small bug that's been annoying me -- we don't treat the null
address as a "bad" coordinate, which means that if MaxMind can't resolve a lat/long for a
particular incoming client request, we send them to null island. Instead, the change
in this commit that checks for lat/long == (0,0) in SortServers will cause a client IP
resolving to the null coordinate to be redirected to a randomly-ordered list of caches.