Origins now advertise themselves as writebackhosts
This is a bandaid fix for now to allow origins to show themselves as
writeback hosts to allow for puts. This helps seperate them from relying
on topology to get this information. This is just a band-aid for now, as
we should make the writebackhost configurable and able to be set or not
set when an origin is spun up.