Add generated files to git
If anyone wants to import the Pelican modules externally in the future
(we're not there with the client -- but would like to be!), then the
generated files need to be part of the release.

This commit adds the generated files and tweaks the parameter struct
generator so the results are stable across multiple runs.

Fixes #387