Move Prometheus configuration until after the web interface has started
Prometheus service discovery will query the web interface for a list
of all known origins and caches.  In the case of the federation mode,
the timing is just right that the service discovery was occurring
before the interface startup completed.

To avoid race conditions, force the Prometheus configuration to occur
after the web UI startup has completed.