Large overhaul of the process and state management
Convert almost all of Pelican to use contexts, allowing any goroutines
launched to be cancelled from an application-level context.  This allows
improved cleanup between unit tests and ensures, on shutdown, that
any temporary resources have been cleaned up.

With this change, we can migrate the module startup to a more uniform
approach, making the federation-in-a-box work better behind a single
port.