Update ns key discovery to follow openid-style lookups
Previously, namespace keys were fetched by hardcoding the issuer jwks relative
to the namespace. Now we follow openid-style lookups by fetching the JSON
located at <registry url>/api/v2.0/registry/metadata/<namespace>/.well-known/namespace-configuration
and following the URL associated with the `jwks_uri` key. Currently this is still the same key path
in all instances that I'm aware of, but in theory this could be expanded to point to other locations.

The architecture of this PR is that a GET request to the namespace-configuration url mentioned above
should dynamically generate the `jwks_uri` key with associated value. The value is used throughout
and assigned the issuer URL in various places where we need to create a token on behalf of the
namespace.