Have cache and origin run on Cache.Port and Origin.Port, respectively
This deprecates the Xrootd.Port configuration and replaces it with
Origin.Port and Cache.Port.  By separating out the configs, we have
some hope of making the cache module part of the "federation in a box"
since the cache will need to be a standalone Xrootd instance.

With this, Server.WebPort and Cache.Port/Origin.Port (as appropriate)
are a hard requirement for the various modules to be enabled, requiring
some unit tests to set these values.