Regenerate the authfile and scitokens.cfg periodically
To replace existing OSDF caches, which rely on X509 authentication that we don't
want to pull into the Pelican codebase, we'll need to merge with their authfiles
(which are periodically updated).  This utilizes the Xrootd maintenance routine
to regenerate authfiles and scitokens.cfg when an update occurs.

Includes unit tests.