Properly parse configurations without audience_json set
The INI library always returns a key when there's a lookup -- if the
key isn't in the section, the returned object is a valid key with an
empty value.  When this empty string was being parsed as JSON, an
error was thrown.

Fixed the conditional to check for a non-empty key and added a short
unit test.