First attempt at a connection brokering service
Introduces a new module in the code, the Broker, which will allow
the cache to initiate a reversed TCP connection to the origin (the
broker will request the origin connect to the cache, then the cache
will treat the TCP client as a server).

The reversal request is initiated by the xrdcl-pelican plugin; to
use this code, an updated plugin may be needed.  On the origin side,
the pelican process will proxy to the local Xrootd using a host certificate
provided by the cache.

The major TODO items (beyond testing and debugging) not implemented
are the creation and verification of tokens to authenticate the
requests to the broker.