Fix origin export options to allow Xrootd.Mount/Origin.NamespacePrefix combo
To specify the local export and namespace prefix when serving an origin, both
`Origin.ExportVolume` or a combination of `Xrootd.Mount` and `Origin.NamespacePrefix`
are valid options. There was a bug that prevented the origin from spinning up in the
case that `Origin.ExportVolume` was not directly specified, or populated via the
command line volume mount flag. This fix allows either combination to work.