Enable OSDF defaults for stashcp/stash_plugin as well
NOTE: This switches the default for the plugin to use the director
which will migrate a large amount of load to the director compared
to what's currently seen.

To revert back to the prior behavior based on OSG Topology, set the
environment variable `STASH_USE_TOPOLOGY` to be non-empty.

Fixes #720