Fixes to allow osdf defaults for stashcp/plugin
The global nature of the metadata configuration caused confusion
when it is used for per-URL director discovery.

Fixes the issue where the discovery url was ignored and was
automatically using the OSDF default.  When the OSDF default was
used, it was causing a test that was supposed to fail to start
succeeding.