Add pelican-server RPM
This is one RPM that provides default configs and SystemD service files for
the origin, cache, registry, and director.  This provides the virtual
dependencies "pelican-origin", "pelican-cache", "pelican-registry", and
"pelican-director" so anyone trying to do `yum install pelican-cache`
will get this RPM.

Due to GoReleaser limitations, an RPM will be built for each arch (even though
there are no arch-specific contents), and I can't make the versions in the
"provides" lines based on the pelican version, I have to hardcode the number.