Add simple log monitoring mechanism for watching XRootD initialization
This adds a simple hook providing callers with the ability to get
callbacks when a logging line matches a regular expression.

Adds such a filter for watching the xrootd daemon for successful startup
(and determining the corresponding port number), speeding up startup as
we no longer need to poll for HTTP server availability.