Rebuild OSG's xrootd RPM from source inside the dev container
This downloads the latest xrootd source RPM from the OSG 23 development
repos, and uses rpmbuild to build it locally for the current
architecture.  The resulting RPMs are put into
/usr/local/src/rpmbuild/RPMS/$arch/* (where $arch is the architecture,
either x86_64, noarch, or aarch64) and a yum repo is built from it.
The repo has highest priority so the locally built RPMs are picked over
what's in EPEL.

To save time and space, the xrootd-compat libraries are not built, and
neither are the docs.  Even with that, though, it's a good 20 minutes to
build on ARM.  Caching the build results might be a good idea.