Remove bucket requirement for S3 origins
It was a mistake to require that S3 origins be configured with a bucket. In
the case of AWS public data, for example, it would be nice to have a single
origin that exports all public buckets from `s3.amazonaws.com`. Under the old
scheme, this would require an origin per bucket -- whoops!