Failure to get metadata is now retryable error
Adds and error to config so that if there is a timeout getting the
metadata response, we mark as retryable and exit(11). In addition, found
that we were marking things as retryable even when the client has no
errors in its checked error stack array, so check for that now and
return false for retryable if that array is empty.