Remove globals in the error handling
The error handling in the client relied on a global stack of errors.
This prevents us from using the client as a library (as multiple concurrent
calls would mess up the global state).

Instead, switch all the call sites to use a single returned error object,
`TransferErrors`, that might hold a list of failures.