Convert the client module into a library
This removes the remaining globals from the client module and converts
the internals to a non-blocking, channel-based setup.

It introduces the concept of a `TransferEngine` which is an instance of
all the background workers needed by the client.  The engine can have one
or more clients, each of which can submit `TransferJob`s.

Each transfer job is unpacked to one or more concrete file transfers (the
job may recursively copy directories) which are executed by the engine's
worker processes.

The progress bar code is removed from the guts of the client and now part
of the command line utilities.

The different public functions (`DoGet`, `DoPut`, `DoCopy`) are refactored
to use the new internal constructs.