Added error when discovery-url contains a path
When a discovery-url contains a path (when set in config or with `-f`),
we will throw an error when we call DiscoverFederation on that url. If
it is contained within a `pelican://` url however, we cannot really
catch that specific error since we will assume any paths added to the
discovery-url to be the namespace-prefix and so on. Therefore, I added
a few notes to the documentation alerting that this will fail.

Note: I realize the client-usage docs are out of date but I added this
for when we refresh the docs