Refactor the HTTP test module
When the `file_cache` module will be added, the `launcher` module will
depend on the `file_cache` module which depends on the client.  That
introduces a circular dependency since the tests are in the client module
itself.

This splits the pieces that need the launcher into a separate `test` module
that breaks the dependency; it works as the launcher-integrated pieces only
depend on the public interfaces of the client.