Add support for "resource scopes"
A resource-enabled scope is of the form:

```
authz:/resource
```

(example: `storage.read:/foo`).  These resources are partially ordered
and hierarchical, meaning some scopes "contain" the authorization for
others.

For example, `storage.read:/foo` implies access to `storage.read:/foo/bar`

If a resource is not given, it is assumed to be `/`.