Fix overwrite of URL path via pointer
The endpoint is a *url.URL with all workers during a download using
the same pointer.  Thus, when one worker modified the URL's path, the
modification was used by others.  This caused a race condition where
the wrong contents of the file would be downloaded.

Also fixed up the unit test which (a) tested the wrong results for
size (looking at the sizes of the upload slice instead of download)
and (b) didn't check the contents of the other file.