Allow posix origins to export multiple prefixes
This should be the first step toward multi-export origins. While I didn't get
things sorted out for S3, the fact is the S3 plugin doesn't really support that
yet anyway. For now, this appears to be working with POSIX, which is the majority
of origins anyway.

In particular, the larges addition as I see it is the `GetOriginExports` function,
which allows us to construct a list of OriginExports that can be used in a standard
way elsewhere.