Sort the caches based on responsiveness
If we cannot get a HEAD request through a cache within a second, it's
a strong signal that it won't be responsive later on when we try to
download.  Do simultaneous HEAD queries and sort those without a response
after 1 second to the end.

The intent is to reduce the time cost of a totally unresponsive cache.