Fix bug where project name not set in user-agent
Project name, if set, will now show up in the user-agent header for
requests. This PR also removes some old unused legacy code of the
"payload" object (the only piece of the object still used was the
project name). Also includes a unit test checking that the client
properly sets this header.