Allow recursive uploads/downloads in plugin
Recursive uploads/downloads of directories work with the pelican plugin
now! Tested with condor, it seems downloads work great however some
functionality is not there for uploads when doing a
transfer_output_remap. Going ahead to push this for now since it seems
to be working for us. To do this, run:

stash_plugin osdf:///<some/path/to/directory>?recursive=true <destination>

I decided to hold off on a unit test since it is not much different from
our current recursive uploads/downloads tests and this can be a test ran
in the condor test infastructure in the future.